requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0
matplotlib==3.7.2
Werkzeug==2.3.7
Flask-CORS==4.0.0
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
huggingface-hub==0.17.3
transformers==4.33.2
timm==0.9.7
albumentations==1.3.1
accelerate==0.23.0
safetensors==0.3.3
